requirements-dev.txt
asgiref==3.7.2; python_version >= "3.7"
atomicwrites==1.4.1; python_version >= "3.7" and python_full_version < "3.0.0" and sys_platform == "win32" or sys_platform == "win32" and python_version >= "3.7" and python_full_version >= "3.4.0"
attrs==23.1.0; python_version >= "3.7"
bandit==1.7.5; python_version >= "3.7"
black==23.3.0; python_version >= "3.7"
certifi==2023.5.7; python_version >= "3.7"
charset-normalizer==3.1.0; python_full_version >= "3.7.0" and python_version >= "3.7"
click==8.1.3; python_version >= "3.7"
colorama==0.4.6; python_version >= "3.7" and python_full_version < "3.0.0" and sys_platform == "win32" and platform_system == "Windows" or sys_platform == "win32" and python_version >= "3.7" and python_full_version >= "3.7.0" and platform_system == "Windows"
coverage==5.5; (python_version >= "2.7" and python_full_version < "3.0.0") or (python_full_version >= "3.5.0" and python_version < "4")
dnspython==2.3.0; python_version >= "3.7" and python_version < "4.0" and python_full_version >= "3.6.1"
email-validator==2.0.0.post2; python_version >= "3.7" and python_full_version >= "3.6.1"
fastapi==0.68.2; python_full_version >= "3.6.1"
gitdb==4.0.10; python_version >= "3.7"
gitpython==3.1.31; python_version >= "3.7"
greenlet==2.0.2; python_version >= "3" and python_full_version < "3.0.0" and (platform_machine == "aarch64" or platform_machine == "ppc64le" or platform_machine == "x86_64" or platform_machine == "amd64" or platform_machine == "AMD64" or platform_machine == "win32" or platform_machine == "WIN32") and (python_version >= "2.7" and python_full_version < "3.0.0" or python_full_version >= "3.6.0") and (python_version >= "3.4" and python_full_version < "3.0.0" and python_version < "4.0" or python_version >= "3.4" and python_version < "4.0" and python_full_version >= "3.6.0") or python_version >= "3" and (platform_machine == "aarch64" or platform_machine == "ppc64le" or platform_machine == "x86_64" or platform_machine == "amd64" or platform_machine == "AMD64" or platform_machine == "win32" or platform_machine == "WIN32") and (python_version >= "2.7" and python_full_version < "3.0.0" or python_full_version >= "3.6.0") and python_full_version >= "3.5.0" and (python_version >= "3.4" and python_full_version < "3.0.0" and python_version < "4.0" or python_version >= "3.4" and python_version < "4.0" and python_full_version >= "3.6.0")
grpcio==1.56.0; python_version >= "3.7"
gunicorn==20.1.0; python_version >= "3.5"
h11==0.14.0; python_version >= "3.7"
httplib2==0.19.1
idna==3.4; python_version >= "3.7" and python_full_version >= "3.6.1"
iniconfig==2.0.0; python_version >= "3.7"
logging-formatter-anticrlf==1.2
markdown-it-py==3.0.0; python_full_version >= "3.7.0" and python_version >= "3.8"
mdurl==0.1.2; python_full_version >= "3.7.0" and python_version >= "3.8"
mock==4.0.3; python_version >= "3.6"
mypy-extensions==0.4.4; python_version >= "3.7"
mypy==0.910; python_version >= "3.5"
packaging==23.1; python_version >= "3.7"
pathspec==0.11.1; python_version >= "3.7"
pbr==5.11.1; python_version >= "3.8"
platformdirs==3.8.0; python_version >= "3.7"
pluggy==1.2.0; python_version >= "3.7"
psycopg2-binary==2.9.6; python_version >= "3.6"
py==1.11.0; python_version >= "3.7" and python_full_version < "3.0.0" or python_full_version >= "3.5.0" and python_version >= "3.7"
pydantic==1.10.9; python_version >= "3.7"
pygments==2.15.1; python_full_version >= "3.7.0" and python_version >= "3.7"
pyparsing==2.4.7; python_version >= "2.6" and python_full_version < "3.0.0" or python_full_version >= "3.3.0"
pytest-asyncio==0.15.1; python_version >= "3.6"
pytest-cov==2.12.1; (python_version >= "2.7" and python_full_version < "3.0.0") or (python_full_version >= "3.5.0")
pytest-mock==3.11.1; python_version >= "3.7"
pytest==6.2.5; python_version >= "3.6"
pyyaml==6.0; python_version >= "3.7"
requests==2.31.0; python_version >= "3.7"
rich==13.4.2; python_full_version >= "3.7.0" and python_version >= "3.7"
six==1.16.0; python_version >= "3.4" and python_full_version < "3.0.0" and python_version < "4.0" or python_version >= "3.4" and python_version < "4.0" and python_full_version >= "3.3.0"
smmap==5.0.0; python_version >= "3.7"
sqlalchemy-utils==0.37.9; python_version >= "3.4" and python_version < "4.0"
sqlalchemy==1.4.48; (python_version >= "2.7" and python_full_version < "3.0.0") or (python_full_version >= "3.6.0")
starlette-json==20.9.18; python_full_version >= "3.6.0"
starlette==0.14.2; python_version >= "3.6" and python_full_version >= "3.6.1"
stevedore==5.1.0; python_version >= "3.8"
toml==0.10.2; python_version >= "3.7" and python_full_version < "3.0.0" or python_full_version >= "3.5.0" and python_version >= "3.7"
tomli==2.0.1; python_version < "3.11" and python_version >= "3.7"
typing-extensions==4.7.0; python_version >= "3.7" and python_full_version >= "3.6.1" and python_version < "3.10"
ujson==4.3.0; python_version >= "3.6"
urllib3==2.0.3; python_version >= "3.7"
uvicorn==0.14.0
